finish NewNoteModal refactor, begin react query for incident notes
